. Added a small patch from Martin Nilsson regarding make_tag_attributes . Removed comments about alloca . Code cleanup, now compiles without any warning with gcc -Wall

+++ caudium/trunk/src/cmods/_Caudium/scratchpad.h
@@ -84,7 +84,7 @@ static inline unsigned char *scratchpad_get(size_t wanted_size)
 
         case -2:
           Pike_error("Wanted size (%lu) exceeds the maximum scratchpad size (%lu)\n",
-                     wanted_size, spad->buf_max);
+                     (long unsigned int)wanted_size, (long unsigned int)spad->buf_max);
     }
   }
   return spad->buf;
